pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tango-python"
version = "0.2.0"
description = "Python SDK for the Tango API"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "MakeGov", email = "tango@makegov.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["tango", "api", "government", "contracts", "sdk"]
dependencies = [
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.3.0",
    "pytest-httpx>=0.30.0",
    "pytest-recording>=0.13.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[project.urls]
Homepage = "https://github.com/makegov/tango-python"
Documentation = "https://docs.makegov.com/tango-python"
Repository = "https://github.com/makegov/tango-python"
Issues = "https://github.com/makegov/tango-python/issues"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=tango --cov-report=term-missing"
markers = [
    "integration: Integration tests that may hit external APIs",
    "live: Tests that always use live API (skip cassettes)",
    "cached: Tests that only run with cached responses",
    "slow: Tests that are slow to execute",
]

[tool.coverage.run]
source = ["tango"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.hatch.build.targets.wheel]
packages = ["tango"]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
]
